BlockController: Force speed to increase when score block disappear
Since the increaseBlockSpeed method has a check for the max speed,
it will easily happen that losing a score block to the danger zone
won't actually increase the speed, because the speed is already at
max. By passing true, it will now force the increase (but with a
new max of 10, to not make it go out of control).