Add a High Score view (which replaces the previous Game Over view)
This new high score view can be reached from the main menu as a
regular high score view, and as a game over view at the end of a
game. If the player's score was good enough to get into the high
score list, an input will appear to let the player type his/her
name and submit the score to the server.